headerfs: create new package for flat header files and db indexing
This commit creates a new package `headerfs ` that is to be used for
storing all types of headers downloaded by neutrino. This includes:
block headers, basic filter headers, and extended filter headers.

The HeaderStore type is split into two components: a single flat file
on disk, and a database index into those flat files. The flat files are
updated in unison with the database index, and crash recovery logic has
been implemented to ensure the state of the file files and database are
always in sync.