headerfs: re-write index to primarily track block headers, filter header are meta-index
This commit re-writes the index partially to _only_ track block header
using the primary index within a single bucket. In place of a new
bucket for each header type, we now instead have special index pointers
that track the tip of the three type of header chains: block, regular
filter, extended filter.

This optimization will reduce the number of keys in the database from:
N * 3, to N + 3, where N is the height of the best known main chain.