neutrino: add Dialer and NameResolver args to Config
In this commit, we add two new arguments to the Config struct that will
allow a caller to specify their own custom dialer along with name
resolver. With this setup, callers will be able to proxy *all* traffic
as they wish. The primary motivation for this change is to add support
for Tor over a SOCKS proxy.