blocklogger: refactor block logger to be a generic "entity progress" logger
In this commit, we refactor the blockProgressLogger struct to instead be
a generic "entity progress" logger. The goal here is to eventually also
use this to log the progress of filter header verification, as the new
sync logic causes the filter header syncing to be slightly behind the
regular block header syncing.