blockmanager: modify the cfHandler goroutine to use new newHeadersSignal
In this commit, we modify the existing cfHandler goroutine to use the
new newHeadersSignal signal. Rather than waiting for a single signal, we
can now continually check to see if we can start the cf headers sync
process. Before this commit, upon restart, we would wait longer than
necessary (or possibly never) for the sync to start as we needed to wait
until we got a _new_ header.

With this commit, we can instead be notified immediately when the block
headers are "far enough". We want to do this as the check pointed sync
for the cf headers is faster than the regular sync.