blockmanager: properly detect checkpoint mismatches for cfheaders
In this commit, we fix a bug that would cause us to fail to actually
detect any mismatched checkpoints during the main cf header checkpointed
sync. Before this commit, checkpoint mismatches would go silently
undetected. The primary cause of this is an off-by-one error that will
be fixed in the next commit.