blockmanager: fix off-by-one bug in request generation for checkpointed cfheaders
In this commit, we fix an off by one error in the code for generating
the requests for the set of checkpointed cfheaders requests. Before this
commit, rather than fetching an initial height range of 1-1000,
1000-2000, etc, we would instead fetch ranges of 0-999, 1000-1999, etc.
The issue here is that we only actually need to slide the height range
over for the very first range as we already have the filter header for
the genesis block on disk.

In this commit, we also re-write the header check point fetching code to
ensure that we have a precisely sized slice for requests, rather than
generating a slice for all possible requests and skipping those that we
already have the headers for on disk.