blockmanager: hold newFilterHeadersMtx for entire duration of writeCFHeadersMsg
In this commit, we modify the writeCFHeadersMsg to hold the
newFilterHeadersMtx mutex for the entire duration of the call. We do
this in order to allow callers to achieve a greater degree of
synchronization by holding the newFilterHeadersMtx mutex. With this new
change, when callers obtain the mutex, they can be certain that no
headers are being written to disk at all.