blockmanager: add new SynchronizeFilterHeaders method
In this commit, we add a new method to the blockManager:
SynchronizeFilterHeaders. This method allows a caller to halt all filter
header syncing/writing and obtain a clean snapshot of the current tip of
the filter header chain. This method will be used in order to fix a bug
in the rescan logic that causes subscribers to miss block messages if as
they're subscribing, new blocks are connected.