blockmanager: simplify IsCurrent() by just using syncPeerMutex
In this commit, we simplify the IsCurrent() method to rely on a mutex,
rather than having to be processed within the internal loop of the
blockManager. With this change, we avoid a possible deadlock between the
cfHandler and the blockmanager. Additionally, we ensure all other
occurrences where we access the syncPeer are properly thread safe.