blockmanager: optimize writeCFHeadersMsg to batch read sys calls
In this commit, we optimize and simplify the writeCFHeadersMsg method.
Before this commit, we would initiate a read request for _each_ header,
meaning each call would incur 1k sys calls. With this new commit, we now
instead will only issue a *single* system call for the entire batch of
headers, resulting in a sizeable performance optimization.