blockmanager: continue to request headers via getheaders at top, not getblocks
In this commit, we modify neutrino to no longer fetch headers at tip via
getblocks. Instead, we'll continue to fetch using getheaders. As a
result, we can simplify the code a bit, and make the code path much
clearer. With this change, we can now avoid tripping up the stall
detection if the remote node doesn't respond to our getblocks message.