blockmanager: in cfHandler use block hash not height for fetch cfheaders condition
In this commit, we modify the main loop for fetching uncheckpointed
cfheaders at tip to use the filter and header tip hash rather than the
height. We do this as this is a more accurate check as it's possible we
re-org back to the same height, but the in memory and disk state have
diverged as the hashes differ yet we didn't act as the heights were the
same.