query: if we're unable to find a valid query peer, exit early in queryPeers
In this commit, we fix a bug that would cause us to continually retry
the same peer (in the case of only a single peer) rather than trying to
exit the loop so we can wait for additional peers we can query. This
also fixes a bug that would cause the rescan loop to just pause
indefinitely, which didn't give is an opportunity to handle the case
where the remote peer doesn't have a filter that we need.