query: allow retries of missed queries, add additional logging within queryBatch
In this commit, we modify the queryBatch method to allow retries of
missed queries. Before this commit, if the remote node didn't respond to
our query (if they were mid re-org, or new block), then we would never
retry that query. As a result, our sync would halt, as we would have a
missing gap in our cfheaders history.

To fix this, we remove skipping failed queries all together, as
eventually a peer should arise that's able to answer our query, as we
currently only use this method for fetching checkpointed cfheaders
queries.