test: extend the cf header mismatch cases to include an OP_RETURN script
In this commit, we modify the TestResolveCFHeadersMismatch test cases to
include a block that has an OP_RETURN output. As is, this test will fail
on master since we'll try to check to see if this the output script is
found within the filter. This is incorrect as the script will be absent
even if the filter itself is correctly formatted.