blockmanager: properly handle case where ALL peers serve bad headers
In this commit, we fix a bug that could at times cause a panic if all
the peers we were connected to served bogus filters. Before this commit,
if the key was blank, then we would get a nil pointer, which would be
passed into `writeCFHeaders`, causing a panic at the first check.