rescan: properly use blockConnected channel for filter fetch retries
In this commit, we fix an existing bug that would prevent us from
actually re-fetching filter header that we missed either due to the peer
not responding, us fetching from the peer while it was mid-reorg, or the
peer simply not being synced yet. We also lower the retry interval to
force us to the best chain tip more aggressively.