utxoscanner: cancel pending reqs + result
This commit adds the ability to cancel GetUtxo
requests via two means:
 * a configurable cancel channel
 * automatically when shutting down the utxoscanner

The first is accomplished by selecting on an extra
cancel, which returns ErrGetUtxoCancelled if signaled
before the result has arrived.

The second iterates through all requests which have
been queued but not started, and cancels them all
with ErrGetUtxoCancelled. This should enable to
dependent systems to break out of long-running
queries, especially during shutdown.