rescan: make `r.running` a channel closed on exit
This commit makes the previously atomic variable running into a channel
that is closed the moment the rescan finishes. This let us select on it
in the Update() method, to ensure we are not trying to enqueue an update
onto an exiting rescan.

Previously we risked that a call to update would block on sending on the
updateChan, and when the rescan was stopped, it would stay blocked. Now
we instead return an error on the failed send.