blockmanager: remove queryResponse length check
queryResponses was recently changed from being a slice to a map.
Previously a length check was performed to make sure we didn't index out
of bounds, but after the change to using a map this is no longer
correct.

This would be apparent doing header download from multiple peers, as
responses would come asynchronously, and the headers wouldn't be written
because of this check.