blockmanager: fetch filter checkpoints up to chain checkpoints
This commit makes the fetching of filter checkpoints go up to the hash
of the latest chain checkpoints in case the block headers haven't caught
up that far.

This helps in terms of avoiding doing a lot of filter checkpoint fetches
up to small heights while the block headers are catching up.