multi: replace block notifications with new subscription manager
In this commit, we replace the current block notification system with
the one provided by the recently introduced
blockntfns.SubscriptionManager. Much of the logic is still the same,
this is merely a refactor to obtain a cleaner abstraction. Notifications
are streamed through a single channel, rather than two, in order to
ensure they are processed in their correct order.