neutrino: ensure we add an address before we modify it
In this commit, we fix a lingering bug w.r.t the way we add new
addresses to the addrmgr that were sourced from purely outbound
connections. Before this commit, we would try to add add the services
for an address before it was known to the addrmgr. This would result in
the `SetServices` method being a noop.

In order to fix this, we now ensure that the addrs are added to the
addrmgr with `AddAddresses` before we mark it as a good address, or try
to set its services.