blockmanager: fix OP_RETURN check during filter conflict resolution
In this commit, we fix a bug in the way we attempt to verify the filters
during conflict resolution. A bug was fixed in the `btcutil` library
that would add some OP_RETURN items to the filter incorrectly. This was
remedied in the library, but this check was never updated itself. We put
all libraries in line by skipping _all_ OP_RETURN outputs when we
attempt to verify the integrity of a filter.