chainsync: populate testnet+mainnet filter header meta checkpoints
In this commit, we add known to be valid filter header checkpoint hashes
for mainnet and testnet. These simply allow peers to avoid downloading
invalid filters and banning peers which we already know are running
incorrect versions of the software. For all blocks past this point,
we'll rely on our normal validation to tend towards the correct filter
header chain.