multi: eliminate the waddrmgr dep by redefining the BlockStamp struct
In this commit, we remove the dep within the codebase on btcwallet as a
hole. After the addition of the new module for the `walletdb` package,
we no longer need to depend on the entire project, making our set of
dependencies much smaller. Additionally, after this change, we've
eliminated the circular dependency between btcwallet and neutrino.