headerfs: add baseline benchmarks
As a preparation for memory behavior optimizations, we first add a set
of benchmark tests to establish a baseline against.
The benchmarks determine the speed and memory usage of writing different
sized batches to the bbolt index as well as the random access latency
for retrieving entries from the index.