neutrino: add spent output script verification
Whenever we download a block, we want to make sure that the filter we
had for it was correct. It's trivial to check all output scripts of the
block's transactions were included in the filter. The trickier part is
to verify the scripts of the outputs being spent are also included in
the filter.
By looking at the signature script or witness we can derive the script
that's being spent and from that can make sure they are all included in
the filter.