query: end completed queries in sendTransaction
(*ChainService).sendTransaction always timed out even if each peer did
respond. This is because after receiving a getdata message in response
to the outgoing inv, the checkResponse closure provided to
queryAllPeers neglected to close the peerQuit channel. As a result,
sendTransaction always blocks for the full configured BroadcastTimeout.

This change adds the missing close(peerQuit) when a matching response
(either a getdata or reject) is received.

Note that as before this change, a non-nil error from sendTransaction
does not guarantee that a remote peer actually received the MsgTx that
was queued since the outgoing tx message is handled asynchronously.
To account for the removed delay, the TestNeutrinoSync unit test is
updated with a waitTx helper to ensure the mining node has the sent
transaction prior to requesting a new block.