query: declare TX invalid based on threshold
If we didn't get a reject message from every peer that replied to our
call with an INV message, we didn't look at the responses at all. With
this commit we check if at least 60% (by default) of our responding
peers declared the TX as invalid. If that's the case, we return the
error since the transaction probably will never be accepted in the
future.