wallet: avoid detecting addrs from non-default scopes
This commit fixes a bug where we'd detect funds sent to a non-default
scope address when adding a relevant TX for another output (e.g. the
change from sending to the non-default scope address). This would lead
to not recognizing the spend of that output later on and the output
would remain as a "ghost UTXO".