chain: always add txes to mempool on receiving `rawtx` events
Although we have a mempool polling, there's still a case where a wanted
tx can be missed to notify. Suppose the events happen in the following
order,
1. a relevant tx is received via `rawtx`.
2. an input is subscribed, and,
3. mempool tx polling just finished its last round, meaning it won't
   refresh the mempool for another `TxPollingInterval`, thus this
   relevant tx is not added to the local mempool.

In order to make sure the relevant tx is in mempool, we now need to also
feed it into the mempool when we receive it from `rawtx` queue.