chain: make mempool loading async, use fine grained locking
For nodes with very large mempools, or node running on meager hardware,
loading the mempoool might take some time. To mitigate this, in this
commit, we move to make loading the mempool async. We also start to use
more fine grained locking to avoid holding the internal mutex the entire
time while we're processing the mempool diff.