query+neutrino: create a WorkManager interface
Introduce a `WorkManager` interface that encapsulates the Dispatcher
interface along with `Start` and `Stop` methods. The previous
`WorkManager` struct is converted to `workManager` which is an
implementation of the new `WorkManager` interface. The `ChainService`
struct is then adapted to hold the new interface instead of the struct.
This will make testing easier in future.