query+blockmanager: add NumRetries optional param
In this commit, the worker manager is adjusted to only retry a request
a certain maximum number of times. This is done so that a smooth
transition can be made from call sights currently using the
neutrino.queryChainServicePeers method to using the query.WorkManager
instead so that the same default behaviour can be expected. In order to
also cater for call sights currently using the query.WorkManager, a
NoRetryMax optional paramater is added that can used to disable the
retry maximum.