chain: add `NewRPCClientWithConfig` to init client using config
This commit adds a new method to init the RPC client. Unlike the
`NewRPCClient`, it now takes a config to allow users specifying
notification handlers.

In the future, `NewRPCClient` should be replaced by this method. For now
we implement it this way so other callsites relying on this method can
stay out of being affected.