chain: update filters synchronously
Previously, when adding filters to be monitored, the items are added to
the internal maps asynchronously. This means when a block is connected,
or a tx is broadcast, we have a small chance of missing the monitored
filters, causing the wallet to report an incorrect balance. This is now
fixed by making sure we always update the maps in those calls.